requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
pbr==2.1.0
Penman==1.1.0
progress==1.5
PyDelphin==1.5.1
requests==2.26.0
six==1.16.0
urllib3==1.26.7
wxconv==1.0.0.0
beautifulsoup4~=4.10.0
googletrans~=3.1.0a0
nltk~=3.6.3